Refactor the utils package to other area
Move cookieSession functions to cookie session store
& align the double implementation of SecretBytes to be
united and housed under encryption
